+++ src/aio.c
@@ -211,25 +211,6 @@ static void raio_complete_signal(void *arg) {
 
 }
 
-static void request_complete_signal(void *arg) {
-
-  nano_aio *raio = (nano_aio *) arg;
-  nano_aio *saio = (nano_aio *) raio->next;
-  nano_cv *ncv = (nano_cv *) saio->next;
-  nng_cv *cv = ncv->cv;
-  nng_mtx *mtx = ncv->mtx;
-
-  const int res = nng_aio_result(raio->aio);
-  if (res == 0)
-    raio->data = nng_aio_get_msg(raio->aio);
-
-  nng_mtx_lock(mtx);
-  raio->result = res - !res;
-  ncv->condition++;
-  nng_cv_wake(cv);
-  nng_mtx_unlock(mtx);
-
-}
 
 static void raio_invoke_cb(void *arg) {
   SEXP call, context, data, cb = (SEXP) arg;
@@ -246,21 +227,21 @@ static void raio_invoke_cb(void *arg) {
   UNPROTECT(3);
 }
 
-static void raio_complete_cb(void *arg) {
+static void request_complete(void *arg) {
 
   nano_aio *raio = (nano_aio *) arg;
+  nano_aio *saio = (nano_aio *) raio->next;
   const int res = nng_aio_result(raio->aio);
   if (res == 0)
     raio->data = nng_aio_get_msg(raio->aio);
   raio->result = res - !res;
 
-  nano_aio *saio = (nano_aio *) raio->next;
   if (saio->data != NULL)
     later2(raio_invoke_cb, saio->data, 0);
 
 }
 
-static void request_complete_cb(void *arg) {
+static void request_complete_signal(void *arg) {
 
   nano_aio *raio = (nano_aio *) arg;
   nano_aio *saio = (nano_aio *) raio->next;
@@ -1205,7 +1186,7 @@ SEXP rnng_ncurl_session_close(SEXP session) {
 
 SEXP rnng_request_impl(const SEXP con, const SEXP data, const SEXP sendmode,
                        const SEXP recvmode, const SEXP timeout, const SEXP clo,
-                       nano_cv *ncv, const int promises) {
+                       nano_cv *ncv) {
 
   const nng_duration dur = timeout == R_NilValue ? NNG_DURATION_DEFAULT : (nng_duration) Rf_asInteger(timeout);
   const int mod = nano_matcharg(recvmode);
@@ -1247,11 +1228,7 @@ SEXP rnng_request_impl(const SEXP con, const SEXP data, const SEXP sendmode,
   raio->mode = mod;
   raio->next = saio;
 
-  if ((xc = nng_aio_alloc(&raio->aio,
-                          promises ?
-                            (signal ? request_complete_cb : raio_complete_cb) :
-                            (signal ? request_complete_signal : raio_complete),
-                          raio)))
+  if ((xc = nng_aio_alloc(&raio->aio, signal ? request_complete_signal : request_complete, raio)))
     goto exitlevel2;
 
   nng_aio_set_timeout(raio->aio, dur);
@@ -1289,7 +1266,7 @@ SEXP rnng_request(SEXP con, SEXP data, SEXP sendmode, SEXP recvmode, SEXP timeou
   if (R_ExternalPtrTag(con) != nano_ContextSymbol)
     Rf_error("'con' is not a valid Context");
 
-  return rnng_request_impl(con, data, sendmode, recvmode, timeout, clo, NULL, 0);
+  return rnng_request_impl(con, data, sendmode, recvmode, timeout, clo, NULL);
 
 }
 
@@ -1301,18 +1278,7 @@ SEXP rnng_request_signal(SEXP con, SEXP data, SEXP cvar, SEXP sendmode, SEXP rec
     Rf_error("'cv' is not a valid Condition Variable");
   nano_cv *ncv = (nano_cv *) R_ExternalPtrAddr(cvar);
 
-  return rnng_request_impl(con, data, sendmode, recvmode, timeout, clo, ncv, 0);
-
-}
-
-SEXP rnng_request_promise(SEXP con, SEXP data, SEXP cvar, SEXP sendmode, SEXP recvmode, SEXP timeout, SEXP clo) {
-
-  if (R_ExternalPtrTag(con) != nano_ContextSymbol)
-    Rf_error("'con' is not a valid Context");
-
-  nano_cv *ncv = R_ExternalPtrTag(cvar) == nano_CvSymbol ? (nano_cv *) R_ExternalPtrAddr(cvar) : NULL;
-
-  return rnng_request_impl(con, data, sendmode, recvmode, timeout, clo, ncv, 1);
+  return rnng_request_impl(con, data, sendmode, recvmode, timeout, clo, ncv);
 
 }